Move all audit views and mixins to the new app
