+++ anvil_consortium_manager/auditor/tests/test_viewmixins.py
@@ -1,0 +1,12 @@
+from django.core.exceptions import ImproperlyConfigured
+from django.test import TestCase
+
+from .. import viewmixins
+
+
+class AnVILAuditMixinTest(TestCase):
+    """ManagedGroupGraphMixin tests that aren't covered elsewhere."""
+
+    def test_run_audit_not_implemented(self):
+        with self.assertRaises(ImproperlyConfigured):
+            viewmixins.AnVILAuditMixin().run_audit()
